refactor: move  from_archive into pop_common
