+++ crates/pop-common/src/errors.rs
@@ -12,6 +12,9 @@ pub enum Error {
 	#[error("a git error occurred: {0}")]
 	Git(String),
 
+	#[error("HTTP error: {0}")]
+	HttpError(#[from] reqwest::Error),
+
 	#[error("IO error: {0}")]
 	IO(#[from] std::io::Error),
