+++ crates/pop-parachains/src/up/mod.rs
@@ -3,7 +3,7 @@
 use crate::errors::Error;
 use glob::glob;
 use indexmap::IndexMap;
-use pop_common::git::GitHub;
+use pop_common::{git::GitHub, Status};
 use sourcing::{GitHub::*, Source, Source::*};
 use std::{
 	fmt::Debug,
@@ -783,17 +783,6 @@ impl Repository {
 
 		Ok(Self { url, reference, package })
 	}
-}
-
-/// Trait for observing status updates.
-pub trait Status {
-	/// Update the observer with the provided `status`.
-	fn update(&self, status: &str);
-}
-
-impl Status for () {
-	// no-op: status updates are ignored
-	fn update(&self, _: &str) {}
 }
 
 /// Attempts to resolve the package manifest from the specified path.
